backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-multipart
python-dotenv
torch
transformers
sentence-transformers
faiss-cpu
pillow
numpy
scikit-learn
pydantic-settings
imagehash
pytest
httpx
